feat(api): move target profile repository update to its BC
