+++ api/tests/integration/infrastructure/repositories/target-profile-repository_test.js
@@ -1,6 +1,5 @@
 import _ from 'lodash';
 
-import { ObjectValidationError } from '../../../../lib/domain/errors.js';
 import { TargetProfile } from '../../../../lib/domain/models/index.js';
 import { DomainTransaction } from '../../../../lib/infrastructure/DomainTransaction.js';
 import * as targetProfileRepository from '../../../../lib/infrastructure/repositories/target-profile-repository.js';
@@ -209,105 +208,6 @@ describe('Integration | Repository | Target-profile', function () {
         _.pick(item, ['id', 'isPublic', 'name', 'organizationId', 'outdated']),
       );
       expect(foundTargetProfilesAttributes).to.deep.equal([]);
-    });
-  });
-
-  describe('#update', function () {
-    it('should update the target profile name', async function () {
-      // given
-      const targetProfile = databaseBuilder.factory.buildTargetProfile({ name: 'Arthur' });
-      await databaseBuilder.commit();
-
-      // when
-      targetProfile.name = 'Karam';
-      await targetProfileRepository.update(targetProfile);
-
-      // then
-      const { name } = await knex('target-profiles').select('name').where('id', targetProfile.id).first();
-      expect(name).to.equal(targetProfile.name);
-    });
-
-    it('should update the target profile description', async function () {
-      // given
-      const targetProfile = databaseBuilder.factory.buildTargetProfile();
-      await databaseBuilder.commit();
-
-      // when
-      targetProfile.description = 'Je change la description';
-      await targetProfileRepository.update(targetProfile);
-
-      // then
-      const { description } = await knex('target-profiles').select('description').where('id', targetProfile.id).first();
-      expect(description).to.equal(targetProfile.description);
-    });
-
-    it('should update the target profile comment', async function () {
-      // given
-      const targetProfile = databaseBuilder.factory.buildTargetProfile();
-      await databaseBuilder.commit();
-
-      // when
-      targetProfile.comment = 'Je change le commentaire';
-      await targetProfileRepository.update(targetProfile);
-
-      // then
-      const { comment } = await knex('target-profiles').select('comment').where('id', targetProfile.id).first();
-      expect(comment).to.equal(targetProfile.comment);
-    });
-
-    it('should outdate the target profile', async function () {
-      // given
-      const targetProfile = databaseBuilder.factory.buildTargetProfile({ outdated: true });
-      await databaseBuilder.commit();
-
-      // when
-      targetProfile.outdate = true;
-      await targetProfileRepository.update(targetProfile);
-
-      // then
-      const { outdated } = await knex('target-profiles').select('outdated').where('id', targetProfile.id).first();
-      expect(outdated).to.equal(targetProfile.outdated);
-    });
-
-    it('should update and return the target profile "isSimplifiedAccess" attribute', async function () {
-      // given
-      const targetProfile = databaseBuilder.factory.buildTargetProfile({ isSimplifiedAccess: false });
-      await databaseBuilder.commit();
-
-      // when
-      targetProfile.isSimplifiedAccess = true;
-      const result = await targetProfileRepository.update(targetProfile);
-
-      // then
-      expect(result).to.be.instanceOf(TargetProfile);
-      expect(result.isSimplifiedAccess).to.equal(true);
-    });
-
-    it('should not update the target profile and throw an error while id not existing', async function () {
-      // given
-      const targetProfile = databaseBuilder.factory.buildTargetProfile({ name: 'Arthur' });
-      await databaseBuilder.commit();
-
-      // when
-      targetProfile.id = 999999;
-      targetProfile.name = 'Karam';
-      const error = await catchErr(targetProfileRepository.update)(targetProfile);
-
-      // then
-      expect(error).to.be.instanceOf(NotFoundError);
-    });
-
-    it('should not update the target profile name for an database error', async function () {
-      // given
-      const targetProfile = databaseBuilder.factory.buildTargetProfile({ name: 'Arthur' });
-      await databaseBuilder.commit();
-
-      // when
-      targetProfile.name = null;
-      const error = await catchErr(targetProfileRepository.update)(targetProfile);
-
-      // then
-      expect(error).to.be.instanceOf(ObjectValidationError);
     });
   });
