+++ api/tests/prescription/target-profile/integration/infrastructure/repositories/target-profile-administration-repository_test.js
@@ -1,0 +1,106 @@
+import { ObjectValidationError } from '../../../../../../lib/domain/errors.js';
+import { TargetProfile } from '../../../../../../lib/domain/models/index.js';
+import * as targetProfileAdministrationRepository from '../../../../../../src/prescription/target-profile/infrastructure/repositories/target-profile-administration-repository.js';
+import { NotFoundError } from '../../../../../../src/shared/domain/errors.js';
+import { catchErr, databaseBuilder, expect, knex } from '../../../../../test-helper.js';
+
+describe('Integration | Repository | Target-profile', function () {
+  describe('#update', function () {
+    it('should update the target profile name', async function () {
+      // given
+      const targetProfile = databaseBuilder.factory.buildTargetProfile({ name: 'Arthur' });
+      await databaseBuilder.commit();
+
+      // when
+      targetProfile.name = 'Karam';
+      await targetProfileAdministrationRepository.update(targetProfile);
+
+      // then
+      const { name } = await knex('target-profiles').select('name').where('id', targetProfile.id).first();
+      expect(name).to.equal(targetProfile.name);
+    });
+
+    it('should update the target profile description', async function () {
+      // given
+      const targetProfile = databaseBuilder.factory.buildTargetProfile();
+      await databaseBuilder.commit();
+
+      // when
+      targetProfile.description = 'Je change la description';
+      await targetProfileAdministrationRepository.update(targetProfile);
+
+      // then
+      const { description } = await knex('target-profiles').select('description').where('id', targetProfile.id).first();
+      expect(description).to.equal(targetProfile.description);
+    });
+
+    it('should update the target profile comment', async function () {
+      // given
+      const targetProfile = databaseBuilder.factory.buildTargetProfile();
+      await databaseBuilder.commit();
+
+      // when
+      targetProfile.comment = 'Je change le commentaire';
+      await targetProfileAdministrationRepository.update(targetProfile);
+
+      // then
+      const { comment } = await knex('target-profiles').select('comment').where('id', targetProfile.id).first();
+      expect(comment).to.equal(targetProfile.comment);
+    });
+
+    it('should outdate the target profile', async function () {
+      // given
+      const targetProfile = databaseBuilder.factory.buildTargetProfile({ outdated: true });
+      await databaseBuilder.commit();
+
+      // when
+      targetProfile.outdate = true;
+      await targetProfileAdministrationRepository.update(targetProfile);
+
+      // then
+      const { outdated } = await knex('target-profiles').select('outdated').where('id', targetProfile.id).first();
+      expect(outdated).to.equal(targetProfile.outdated);
+    });
+
+    it('should update and return the target profile "isSimplifiedAccess" attribute', async function () {
+      // given
+      const targetProfile = databaseBuilder.factory.buildTargetProfile({ isSimplifiedAccess: false });
+      await databaseBuilder.commit();
+
+      // when
+      targetProfile.isSimplifiedAccess = true;
+      const result = await targetProfileAdministrationRepository.update(targetProfile);
+
+      // then
+      expect(result).to.be.instanceOf(TargetProfile);
+      expect(result.isSimplifiedAccess).to.equal(true);
+    });
+
+    it('should not update the target profile and throw an error while id not existing', async function () {
+      // given
+      const targetProfile = databaseBuilder.factory.buildTargetProfile({ name: 'Arthur' });
+      await databaseBuilder.commit();
+
+      // when
+      targetProfile.id = 999999;
+      targetProfile.name = 'Karam';
+      const error = await catchErr(targetProfileAdministrationRepository.update)(targetProfile);
+
+      // then
+      expect(error).to.be.instanceOf(NotFoundError);
+    });
+
+    it('should not update the target profile name for an database error', async function () {
+      // given
+      const targetProfile = databaseBuilder.factory.buildTargetProfile({ name: 'Arthur' });
+      await databaseBuilder.commit();
+
+      // when
+      targetProfile.name = null;
+      const error = await catchErr(targetProfileAdministrationRepository.update)(targetProfile);
+
+      // then
+      expect(error).to.be.instanceOf(ObjectValidationError);
+    });
+  });
+});
